ml-training/requirements.txt
torch==2.1.0
torchvision==0.16.0
transformers==4.36.0
datasets==2.15.0
pillow==10.1.0
numpy==1.24.3
pandas==2.1.3
scikit-learn==1.3.2
albumentations==1.3.1
opencv-python==4.8.1.78
tqdm==4.66.1
matplotlib==3.8.2
seaborn==0.13.0
huggingface-hub==0.19.4
wandb==0.16.0
accelerate==0.25.0